+++ example/swipeable_button_example_1/lib/main.dart
@@ -27,16 +27,63 @@ class SwipeableButtonExample1HomePage extends StatelessWidget {
     ),
     body: Center(
       child: SizedBox(
-        width: 200.0,
-        child: SwipeableButton(
-          label: const Center(child: Text("Swipe me!", style: TextStyle(color: Colors.white, fontSize: 16.0))),
-          color: Colors.blue.shade400,
-          thumbColor: Colors.blue.shade800,
-          onSwipe: () {
-            ScaffoldMessenger.of(context).showSnackBar(
-              const SnackBar(content: Text("Thank you for swiping!"),)
-            );
-          }
+        width: 300.0,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: SwipeableButton.simple(
+                label: const Center(child: Text("Simple blue button", style: TextStyle(color: Colors.white, fontSize: 16.0))),
+                color: Colors.blue.shade400,
+                thumbColor: Colors.blue.shade800,
+                onSwipe: () {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text("Thank you for swiping!"),)
+                  );
+                }
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: SwipeableButton.simple(
+                label: const Center(child: Text("Simple red button", style: TextStyle(color: Colors.white, fontSize: 16.0))),
+                color: Colors.red.shade400,
+                thumbColor: Colors.red.shade800,
+                onSwipe: () {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text("Thank you for swiping!"),)
+                  );
+                }
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: SwipeableButton(
+                label: const Center(child: Text("Red to green thumb", style: TextStyle(color: Colors.white, fontSize: 16.0))),
+                color: Colors.blue.shade400,
+                thumbBuilder: (BuildContext context, double swipedFraction, bool hasSwiped) => DecoratedBox(
+                  decoration: BoxDecoration(color: Color.lerp(Colors.red, Colors.green, swipedFraction)),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.end,
+                    children: <Widget>[
+                      SizedBox(
+                        width: 40.0,
+                        child: hasSwiped
+                          ? const Icon(Icons.check, color: Colors.white)
+                          : const Icon(Icons.chevron_right, color: Colors.white),
+                      ),
+                    ],
+                  ),
+                ),
+                onSwipe: () {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text("Thank you for swiping!"),)
+                  );
+                }
+              ),
+            ),
+          ],
         ),
       ),
     ),