+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.1.0
+
+* The previous `SwipeableButton` can now be built via the named constructor
+  `SwipeableButton.simple`
+
+* The default `SwipeableButton` constructor now takes a `thumbBuilder` function
+  instead of a `thumbColor` parameter, so the thumb widget can be fully
+  customized
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+Initial version
+
+* The button can be triggered either one time only or infinite times.
+
+* The label shown on the button can be any widget.
+
+* The button's border radius, the button color and the thumb's color can be
+  customized.
+
+* Height and minimum thumb width can be specified.